+++ src/dapps/RegisterDomain/RegisterDomain.vue
@@ -42,7 +42,7 @@ import DeedContractAbi from '@/helpers/deedContractAbi';
 import bip39 from 'bip39';
 import * as unit from 'ethjs-unit';
 import * as nameHashPckg from 'eth-ens-namehash';
-import * as uts46 from 'idna-uts46';
+import normalise from '@/helpers/normalise';
 
 const ETH_TLD = '.eth';
 
@@ -214,12 +214,6 @@ export default {
       );
       return registrarAddress;
     },
-    normalise(str) {
-      return uts46.toUnicode(str, {
-        useStd3ASCII: true,
-        transitional: false
-      });
-    },
     async checkDomain() {
       const web3 = this.$store.state.web3;
       this.loading = true;
@@ -279,12 +273,12 @@ export default {
         this.domainNameErr = false;
       }
       try {
-        this.normalise(value);
+        normalise(value);
       } catch (e) {
         this.domainNameErr = true;
         return;
       }
-      this.domainName = this.normalise(value);
+      this.domainName = normalise(value);
     },
     async getMoreInfo(deedOwner) {
       const deedContract = new this.$store.state.web3.eth.Contract(
@@ -352,7 +346,8 @@ export default {
       const revealDate = date.setDate(date.getDate() - 2);
       const raw = {
         from: address,
-        value: type === 'reveal' ? 0 : unit.toWei(this.bidMask, 'ether').toString(),
+        value:
+          type === 'reveal' ? 0 : unit.toWei(this.bidMask, 'ether').toString(),
         to: this.registrarAddress,
         data: contractReference.encodeABI(),
         name: this.domainName,